requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
phonenumbers==8.12.55
django-phonenumber-field==7.0.0
djangorestframework==3.13.1
